Week8/requirements.txt
pandas
faiss-cpu
sentence-transformers
openai
streamlit
